+++ src/methods/set/set.js
@@ -48,6 +48,13 @@ module.exports = {
     let diff = s.minute() - n
     let shift = diff * ms.minute
     s.epoch -= shift
+    // check against a screw-up
+    if (old.hour() != s.hour()) {
+      walkTo(old, {
+        minute: n
+      })
+      return old.epoch
+    }
     confirm(s, old, 'second')
     return s.epoch
   },